reworked dashboard chart container
