+++ client/package.json
@@ -7,6 +7,7 @@
     "@material-ui/core": "^4.2.0",
     "axios": "^0.19.0",
     "react": "^16.8.6",
+    "react-c3js": "^0.1.20",
     "react-dom": "^16.8.6",
     "react-router-dom": "^5.0.1",
     "react-scripts": "3.0.1"
