+++ package.json
@@ -29,6 +29,8 @@
     "passport": "^0.4.0",
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
-    "passport-local-mongoose": "^5.0.1"
-  }
+    "passport-local-mongoose": "^5.0.1",
+    "react-router-dom": "^5.0.1"
+  },
+  "devDependencies": {}
 }
